backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Alisson",email = "alissonmarcos249@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "fastapi[standard] (>=0.135.1,<0.136.0)",
    "sqlalchemy[asyncio] (>=2.0.49,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "aiosqlite (>=0.22.1,<0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.15.7,<0.16.0)",
    "typos (>=1.44.0,<2.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "freezegun (>=1.5.5,<2.0.0)"
]

[tool.coverage.run]
core = "ctrace"
concurrency = ["thread", "greenlet"]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.taskipy.tasks]
pre_lint = 'typos'
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev backend/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=backend -vv'
post_test = 'coverage html'